pyproject.toml
[project]
name = "stable-diffusion-drawbridge-api"
version = "1.3"
description = "使用标准A1111-SDWEBUI API来调用各种绘画后端的多功能负载均衡API"
authors = [
    {name = "DiaoDaiaChan", email = "diaodaiachan@qq.com"},
]
dependencies = [
    "aiohttp",
    "aiofiles",
    "Pillow>=10.4.0",
    "civitai-py",
    "civitai",
    "colorama",
    "fastapi",
    "fal-client",
    "httpx",
    "piexif",
    "pydantic",
    "PyYAML",
    "redis",
    "replicate",
    "tqdm",
    "uvicorn",
    "gradio"
]

[tool.pdm]
exclude = ["./DrawBridgeAPI/config.yaml", "./DrawBridgeAPI/*.log", "./DrawBridgeAPI/saved_images"]

requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

